+++ packages/jarun/index.ts
@@ -4,5 +4,4 @@ export * from "./JarunPromise";
 export * from "./JarunTestProvision";
 export * from "./JarunTestRunner";
 export * from "./util";
-
 export * from "^jarunc";