+++ packages/jatec/index.ts
@@ -1,3 +1,4 @@
 export * from "./types";
 export * from "./typesTest";
+export * from "./typesFramework";
 export * from "./util";