+++ packages/jatec/typesFramework.ts
@@ -1,0 +1,6 @@
+import { TestResult } from "./internal";
+
+export type TestRunner = {
+  runTest: (id: string, absTestFile: string) => Promise<TestResult>;
+  kill: () => Promise<void>;
+};